+++ ChangeLog.md
@@ -4,6 +4,7 @@
 * Vitalité surnaturelle du barbare
 * Riposte du guerrier
 * Nuée de criquets (Druide Prestige)
+* Increvable (Rôdeur, Survie rang 4)
 * Intervention divine du prêtre
 
 ### Autres améliorations